refactor(admin): split settings.ts into per-feature route modules
settings.ts had grown to 938 lines mixing unrelated concerns. Continue
the existing settings-*.ts extraction pattern, moving each remaining
handler group into its own module and leaving settings.ts as a thin
route aggregator (938 -> 102 lines).

New modules:
- settings-page.ts        GET handlers + page-state gathering
- settings-password.ts    change-password handler
- settings-stripe.ts      Stripe credentials/webhook/test
- settings-square.ts      Square credentials/webhook/test
- settings-email.ts       email provider config + test send
- settings-header-image.ts header image upload/delete
- settings-superuser.ts   superuser recovery enable/decline
- settings-general.ts     small field/toggle handlers, column order, reset

Also:
- move shared getWebhookUrl into settings-helpers.ts (used by page + stripe)
- drop dead processSecretField/SecretFieldResult re-exports from settings.ts
  (every consumer imports them straight from settings-helpers.ts)

Pure move, no behaviour change. Route keys are unchanged and the suite
drives handlers through the HTTP router, so coverage stays at 100%.
