+++ src/features/admin/settings-helpers.ts
@@ -11,6 +11,7 @@
 
 import { type AuthSession, OWNER_FORM, withAuth } from "#routes/auth.ts";
 import { errorRedirect, jsonResponse, redirect } from "#routes/response.ts";
+import { getEffectiveDomain } from "#shared/config.ts";
 import { logActivity } from "#shared/db/activityLog.ts";
 import { isMaskSentinel } from "#shared/db/settings.ts";
 import type { FormParams } from "#shared/form-data.ts";
@@ -67,6 +68,11 @@ const testRoute =
   (testFn: () => Promise<unknown>) =>
   (request: Request): Promise<Response> =>
     withAuth(request, OWNER_FORM, async () => jsonResponse(await testFn()));
+
+/** Build the payment webhook URL from the configured domain.
+ * Shared by the settings page (display) and the Stripe handler (setup). */
+const getWebhookUrl = (): string =>
+  `https://${getEffectiveDomain()}/payment/webhook`;
 
 /** Run an optional async validator; return error response or null */
 const runValidate = async <T>(
@@ -268,6 +274,7 @@ export {
   advancedSettingsRoute,
   clearableFieldHandler,
   createSettingsHandler,
+  getWebhookUrl,
   processSecretField,
   secretFieldHandler,
   settingsClearable,
